pyproject.toml
[tool.poetry]
name = "rbac project"
version = "0.1.0"
description = ""
authors = ["Ojo Peter <ojopeter36@gmail.com>"]
license = "MIT"


[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "1.0.1"
aiosmtplib = "2.0.2"
annotated-types = "0.6.0"
anyio = "4.2.0"
arq = "0.25.0"
async-timeout = "4.0.3"
blinker = "1.7.0"
certifi = "2023.11.17"
click = "8.1.7"
colorama = "0.4.6"
dnspython = "2.5.0"
ecdsa = "0.18.0"
email-validator = "2.1.0.post1"

fastapi-mail = "1.4.1"
greenlet = "2.0.2"
gunicorn = "21.2.0"
h11 = "0.14.0"
hiredis = "2.3.2"
httpcore = "1.0.2"
httptools = "0.6.1"
httpx = "0.25.2"
idna = "3.6"
iniconfig = "2.0.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.3"
mako = "1.3.2"
markupsafe = "2.1.4"
orjson = "3.9.12"
packaging = "23.2"
passlib = "1.7.4"
pluggy = "1.4.0"
psycopg2-binary = "2.9.9"
pyasn1 = "0.5.1"
pydantic = "2.6.0"
pydantic-extra-types = "2.5.0"
pydantic-settings = "2.1.0"
pydantic-core = "2.16.1"
pytest = "7.4.4"
python-jose = "3.3.0"
python-multipart = "0.0.6"
pyyaml = "6.0.1"
redis = "5.0.1"
rsa = "4.9"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.25"
sqlalchemy-utils = "0.41.1"
typing-extensions = "4.9.0"
ujson = "5.9.0"
uuid = "1.30"
uvicorn = "0.23.2"
watchfiles = "0.21.0"
websockets = "12.0"
fastapi = {extras = ["all"], version = "^0.109.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"